Abstract away parts of our wuibutton tag so we can use them later elsewhere.
